Updated reference tests
