+++ tests/reference/asr-bindc_03-95dbba7.json
@@ -2,12 +2,12 @@
     "basename": "asr-bindc_03-95dbba7",
     "cmd": "lpython --show-asr --no-color {infile} -o {outfile}",
     "infile": "tests/errors/bindc_03.py",
-    "infile_hash": "966305cef9890644f613e81f7405e694af6fa13e1b06f0e3e08782bb",
+    "infile_hash": "68e9bc4105ff94df854ee89163425cdc8fd26bcb5615938f6993985c",
     "outfile": null,
     "outfile_hash": null,
     "stdout": null,
     "stdout_hash": null,
     "stderr": "asr-bindc_03-95dbba7.stderr",
-    "stderr_hash": "bd49feaada7484eafec316056295c58526a05d426d1a28f1bb5d8b93",
+    "stderr_hash": "c883db0632bbc273091e912f2e87f9530b662185953c2869bc96ed8f",
     "returncode": 2
 }
