+++ tests/reference/asr-test_numpy_04-ecbb614.json
@@ -6,7 +6,7 @@
     "outfile": null,
     "outfile_hash": null,
     "stdout": "asr-test_numpy_04-ecbb614.stdout",
-    "stdout_hash": "0ed6c5b385b2ea5a829073d375bbdb87b6c06bdc27c6e16d94f10f7d",
+    "stdout_hash": "497129100c171b5c0b1e958275aabc6d9b53e0c06d80f6568ad2d5e9",
     "stderr": null,
     "stderr_hash": null,
     "returncode": 0
