+++ tests/reference/asr-test_numpy_04-ecbb614.stdout
@@ -163,7 +163,11 @@
                                                 3.000000
                                                 (Real 8)
                                             )]
-                                            (Real 8)
+                                            (Array
+                                                (Real 8)
+                                                [(()
+                                                ())]
+                                            )
                                             RowMajor
                                         )
                                         ()
@@ -347,7 +351,11 @@
                                             [(IntegerConstant 1 (Integer 4))
                                             (IntegerConstant 2 (Integer 4))
                                             (IntegerConstant 3 (Integer 4))]
-                                            (Integer 4)
+                                            (Array
+                                                (Integer 4)
+                                                [(()
+                                                ())]
+                                            )
                                             RowMajor
                                         )
                                         ()
