+++ tests/reference/asr-test_pointer_types-1bf0d01.json
@@ -8,6 +8,6 @@
     "stdout": null,
     "stdout_hash": null,
     "stderr": "asr-test_pointer_types-1bf0d01.stderr",
-    "stderr_hash": "4f3b9512978dfb5d22367aa7b2e4fd20768de78530bf4d83a52fa881",
+    "stderr_hash": "3fcd692a6b79b6b3f22fc7a2957df19219fc817446aa6d46f47d049f",
     "returncode": 2
 }
