+++ tests/tests.toml
@@ -1080,6 +1080,10 @@ filename = "errors/bindc_03.py"
 asr = true
 
 [[test]]
+filename = "errors/bindc_04.py"
+asr = true
+
+[[test]]
 filename = "errors/cptr_01.py"
 asr = true
